[Chore] #11 - 파일 구조 개선 (types 분리)
