+++ lib/api/types/common.ts
@@ -1,0 +1,18 @@
+// =====================
+// Common Types
+// =====================
+export interface PaginatedResponse<T> {
+  content: T[]
+  totalElements: number
+  totalPages: number
+  page: number
+  size: number
+  hasNext: boolean
+  hasPrevious: boolean
+}
+
+export interface PaginationParams {
+  page?: number
+  size?: number
+  sort?: string
+}
