+++ lib/api/types/index.ts
@@ -1,0 +1,19 @@
+// =====================
+// Unified Type Exports
+// =====================
+// Export all types for backward compatibility with existing imports
+// This allows imports like: import type { Product, Cart } from './types'
+
+export * from './common'
+export * from './auth'
+export * from './product'
+export * from './cart'
+export * from './order'
+export * from './farm'
+export * from './experience'
+export * from './review'
+export * from './payment'
+export * from './delivery'
+export * from './search'
+export * from './notification'
+export * from './seller'
